refactor more logs
