+++ androguard/core/androconf.py
@@ -167,38 +167,6 @@ def save_colors():
     return c
 
 
-def long2int(l):
-    if l > 0x7fffffff:
-        l = (0x7fffffff & l) - 0x80000000
-    return l
-
-
-def long2str(l):
-    """Convert an integer to a string."""
-    if type(l) not in (types.IntType, types.LongType):
-        raise ValueError('the input must be an integer')
-
-    if l < 0:
-        raise ValueError('the input must be greater than 0')
-    s = ''
-    while l:
-        s = s + chr(l & 255)
-        l >>= 8
-
-    return s
-
-
-def str2long(s):
-    """Convert a string to a long integer."""
-    if type(s) not in (types.StringType, types.UnicodeType):
-        raise ValueError('the input must be a string')
-
-    l = 0
-    for i in s:
-        l <<= 8
-        l |= ord(i)
-
-    return l
 
 
 def is_android(filename):
